backend/requirements.txt
# FastAPI Backend - Antambahoaka Connect
# Installation: pip install -r requirements.txt

# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Email
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0

# CORS & Middleware
# FastAPI has built-in CORS support via fastapi.middleware.cors

# Utils
python-dateutil==2.8.2
